API

PosteriorDB.PosteriorDB

PosteriorDB.AbstractImplementation

PosteriorDB.Dataset

PosteriorDB.Model

PosteriorDB.Posterior

PosteriorDB.PosteriorDatabase

PosteriorDB.ReferencePosterior

PosteriorDB.StanModelImplementation

PosteriorDB.database

PosteriorDB.database

PosteriorDB.dataset

PosteriorDB.dataset

PosteriorDB.dataset_names

PosteriorDB.implementation

PosteriorDB.implementation_names

PosteriorDB.info

PosteriorDB.load

PosteriorDB.load

PosteriorDB.load

PosteriorDB.model

PosteriorDB.model

PosteriorDB.model_names

PosteriorDB.name

PosteriorDB.path

PosteriorDB.path

PosteriorDB.path

PosteriorDB.path

PosteriorDB.posterior

PosteriorDB.posterior_names

PosteriorDB.recursive_stack

PosteriorDB.reference_posterior

PosteriorDB.reference_posterior

PosteriorDB.PosteriorDB — Module

PosteriorDB

Functionality for working with instances of a posterior database.

See https://github.com/stan-dev/posteriordb for more information.

PosteriorDB.AbstractImplementation — Type

AbstractImplementation

Abstract type for implementations of a model.

Subtypes should implement the methods load and path.

PosteriorDB.Dataset — Type

Dataset

Pointer to a dataset in a PosteriorDatabase.

See also: dataset, dataset_names

PosteriorDB.Model — Type

Model

Pointer to a model in a PosteriorDatabase.

See also: model, model_names

PosteriorDB.Posterior — Type

Posterior

Pointer to a posterior in a PosteriorDatabase.

See also: posterior, posterior_names

PosteriorDB.PosteriorDatabase — Type

PosteriorDatabase

A posterior database object using the files at path.

See also: database, posterior_names, model_names, dataset_names

PosteriorDB.ReferencePosterior — Type

ReferencePosterior

Pointer to a reference posterior in a PosteriorDatabase.

See also: Posterior, reference_posterior, posterior

PosteriorDB.StanModelImplementation — Type

StanModelImplementation

Implementation of a model using Stan.

See also: implementation, load

PosteriorDB.database — Function

database([path::String]) -> PosteriorDatabase

Return a pointer to the PosteriorDatabase stored at path.

If path is not given, the default posterior database downloaded from a GitHub release is used.

Note

Julia 1.3 or greater is required to use the default posterior database with database().

PosteriorDB.database — Method

database(obj) -> PosteriorDatabase

Return the database containing obj.

obj can be a Posterior, Model, AbstractImplementation, Dataset, or ReferencePosterior.

PosteriorDB.dataset — Method

dataset(db::PosteriorDatabase, name::String) -> Dataset

Return a dataset object for the dataset with name name in the database db.

PosteriorDB.dataset — Method

dataset(posterior::Posterior) -> Dataset

Return the dataset entry of posterior.

PosteriorDB.dataset_names — Method

dataset_names(db::PosteriorDatabase) -> Vector{String}

Return the names of all datasets in the database.

PosteriorDB.implementation — Method

implementation(model::Model, framework::String) -> AbstractImplementation

Return the implementation of the model for the given framework.

PosteriorDB.implementation_names — Method

implementation_names(model::Model) -> Vector{String}

Return the names of frameworks with code for the model.

PosteriorDB.info — Method

info(obj) -> AbstractDict

Return the info dictionary for obj.

obj can be a Posterior, Model, Dataset, or ReferencePosterior.

PosteriorDB.load — Method

load(implementation::AbstractImplementation) -> String

Load the model implementation and return the code as a string.

PosteriorDB.load — Method

load(dataset::Dataset) -> OrderedDict{String,Any}

Load and return the data for dataset.

load(dataset::Dataset, String) -> String

Return the data for dataset as an unparsed JSON string.

PosteriorDB.load — Method

load(rp::ReferencePosterior) -> Vector{OrderedDict{String,Any}}

Load and return the reference draws for the reference posterior rp.

load(rp::ReferencePosterior, String) -> String

Return the reference draws as an unparsed JSON string.

PosteriorDB.model — Method

model(db::PosteriorDatabase, name::String) -> Model

Return a model object for the model with name name in the database db.

PosteriorDB.model — Method

model(posterior::Posterior) -> Model

Return the model entry of posterior.

PosteriorDB.model_names — Method

model_names(db::PosteriorDatabase) -> Vector{String}

Return the names of all models in the database.

PosteriorDB.name — Method

name(obj) -> String

Return the name of obj.

obj can be a Posterior, Model, Dataset, or ReferencePosterior.

PosteriorDB.path — Method

path(implementation::AbstractImplementation) -> String

Absolute path to the file containing the model implementation.

PosteriorDB.path — Method

path(dataset::Dataset) -> String

Absolute path to the file containing the model dataset.

PosteriorDB.path — Method

path(db::PosteriorDatabase) -> String

Absolute path to the database db.

PosteriorDB.path — Method

path(rp::ReferencePosterior) -> String

Return the path to the file storing the reference draws for the reference posterior rp.

PosteriorDB.posterior — Method

posterior(db::PosteriorDatabase, name::String) -> Posterior

Return a posterior object for the posterior with name name in the database db.

PosteriorDB.posterior_names — Method

posterior_names(db::PosteriorDatabase) -> Vector{String}

Return the names of all posteriors in the database.

PosteriorDB.recursive_stack — Method

recursive_stack(x::AbstractArray)

If x is an array of arrays, recursively stack into a single array whose dimensions are ordered with dimensions of the innermost container first and outermost last.

PosteriorDB.reference_posterior — Method

reference_posterior(db::PosteriorDatabase, name::String) -> ReferencePosterior

Return a pointer to the reference posterior with name name in the database db.

PosteriorDB.reference_posterior — Method

reference_posterior(posterior::Posterior) -> Union{Nothing,ReferencePosterior}

Return the reference posterior entry of posterior.

If no reference posterior is available, nothing is returned.