API

PosteriorDB.PosteriorDB

PosteriorDB.Dataset

PosteriorDB.Model

PosteriorDB.Posterior

PosteriorDB.PosteriorDatabase

PosteriorDB.ReferencePosterior

PosteriorDB.database

PosteriorDB.database

PosteriorDB.dataset

PosteriorDB.dataset

PosteriorDB.dataset_names

PosteriorDB.implementation

PosteriorDB.implementation_names

PosteriorDB.info

PosteriorDB.load_values

PosteriorDB.load_values

PosteriorDB.model

PosteriorDB.model

PosteriorDB.model_names

PosteriorDB.name

PosteriorDB.posterior

PosteriorDB.posterior_names

PosteriorDB.reference_posterior

PosteriorDB.reference_posterior

PosteriorDB.PosteriorDB — Module

PosteriorDB

Functionality for working with instances of a posterior database.

See https://github.com/stan-dev/posteriordb for more information.

PosteriorDB.Dataset — Type

Dataset

Pointer to a dataset in a PosteriorDatabase.

See also: dataset, dataset_names

PosteriorDB.Model — Type

Model

Pointer to a model in a PosteriorDatabase.

See also: model, model_names

PosteriorDB.Posterior — Type

Posterior

Pointer to a posterior in a PosteriorDatabase.

See also: posterior, posterior_names

PosteriorDB.PosteriorDatabase — Type

PosteriorDatabase

A posterior database object using the files at path.

See also: database, posterior_names, model_names, dataset_names

PosteriorDB.ReferencePosterior — Type

ReferencePosterior

Pointer to a reference posterior in a PosteriorDatabase.

See also: Posterior, reference_posterior, posterior

PosteriorDB.database — Function

database([path::String]) -> PosteriorDatabase

Return a pointer to the PosteriorDatabase stored at path.

If path is not given, the default posterior database downloaded from a GitHub release is used.

PosteriorDB.database — Function

database(obj) -> PosteriorDatabase

Return the database containing obj.

obj can be a Posterior, Model, Dataset, or ReferencePosterior.

PosteriorDB.dataset — Method

dataset(db::PosteriorDatabase, name::String) -> Dataset

Return a dataset object for the dataset with name name in the database db.

PosteriorDB.dataset — Method

dataset(posterior::Posterior) -> Dataset

Return the dataset entry of posterior.

PosteriorDB.dataset_names — Method

dataset_names(db::PosteriorDatabase) -> Vector{String}

Return the names of all datasets in the database.

PosteriorDB.implementation — Method

implementation(model::Model, framework::String) -> String

Return the code for the implementation of model in the framework.

PosteriorDB.implementation_names — Method

implementation_names(model::Model) -> Vector{String}

Return the names of frameworks with code for the model.

PosteriorDB.info — Function

info(obj) -> AbstractDict

Return the info dictionary for obj.

obj can be a Posterior, Model, Dataset, or ReferencePosterior.

PosteriorDB.load_values — Method

load_values(dataset::Dataset)

Load and return the data for dataset.

PosteriorDB.load_values — Method

load_values(rp::ReferencePosterior)

Load and return the reference draws for the reference posterior rp.

PosteriorDB.model — Method

model(db::PosteriorDatabase, name::String) -> Model

Return a model object for the model with name name in the database db.

PosteriorDB.model — Method

model(posterior::Posterior) -> Model

Return the model entry of posterior.

PosteriorDB.model_names — Method

model_names(db::PosteriorDatabase) -> Vector{String}

Return the names of all models in the database.

PosteriorDB.name — Function

name(obj) -> String

Return the name of obj.

obj can be a Posterior, Model, Dataset, or ReferencePosterior.

PosteriorDB.posterior — Method

posterior(db::PosteriorDatabase, name::String) -> Posterior

Return a posterior object for the posterior with name name in the database db.

PosteriorDB.posterior_names — Method

posterior_names(db::PosteriorDatabase) -> Vector{String}

Return the names of all posteriors in the database.

PosteriorDB.reference_posterior — Method

reference_posterior(db::PosteriorDatabase, name::String) -> ReferencePosterior

Return a pointer to the reference posterior with name name in the database db.

PosteriorDB.reference_posterior — Method

reference_posterior(posterior::Posterior) -> Union{Nothing,ReferencePosterior}

Return the reference posterior entry of posterior.

If no reference posterior is available, nothing is returned.